+++ packages/fiori-app-sub-generator/test/unit/utils/common.test.ts
@@ -24,6 +24,8 @@ import {
     getReadMeDataSourceLabel,
     getRequiredOdataVersion
 } from '../../../src/utils/common';
+import { isAppStudio } from '@sap-ux/btp-utils';
+import type { Logger } from '@sap-ux/logger';
 
 const getProjectTypeMock = jest.fn();
 jest.mock('@sap-ux/project-access', () => ({
@@ -33,6 +35,10 @@ jest.mock('@sap-ux/project-access', () => ({
 
 jest.mock('@sap-ux/fiori-tools-settings', () => ({
     writeApplicationInfoSettings: jest.fn()
+}));
+
+jest.mock('@sap-ux/launch-config', () => ({
+    createLaunchConfig: jest.fn()
 }));
 
 jest.mock('fs', () => {
@@ -47,6 +53,8 @@ jest.mock('@sap-ux/btp-utils', () => ({
     ...jest.requireActual('@sap-ux/btp-utils'),
     isAppStudio: jest.fn()
 }));
+
+const isAppStudioMock = isAppStudio as jest.Mock;
 
 jest.mock('@sap-ux/fiori-generator-shared', () => ({
     ...jest.requireActual('@sap-ux/fiori-generator-shared'),
@@ -267,6 +275,7 @@ describe('Test utils', () => {
         });
 
         it('should generate correct launch config for OData v2', async () => {
+            isAppStudioMock.mockReturnValue(true);
             // Call the function under test
             await generateLaunchConfig(
                 {
@@ -279,7 +288,7 @@ describe('Test utils', () => {
                 },
                 editor,
                 mockVsCode,
-                undefined,
+                {} as Logger,
                 false
             );
 
@@ -298,14 +307,16 @@ describe('Test utils', () => {
             const expectedFioriOptions = {
                 name: mockProject.name,
                 projectRoot: projectPath,
-                debugOptions: expectedDebugOptions
+                debugOptions: expectedDebugOptions,
+                startFile: undefined
             };
 
-            await createLaunchConfig(projectPath, expectedFioriOptions, editor);
+            expect(createLaunchConfig).toHaveBeenCalledWith(projectPath, expectedFioriOptions, editor, {});
             expect(writeApplicationInfoSettings).toBeCalledWith(projectPath);
         });
 
         it('should generate correct launch config for OData v4', async () => {
+            isAppStudioMock.mockReturnValue(false);
             await generateLaunchConfig(
                 {
                     targetFolder: mockProject.targetFolder,
@@ -313,19 +324,21 @@ describe('Test utils', () => {
                     flpAppId: mockProject.flpAppId,
                     sapClientParam: buildSapClientParam('001'),
                     odataVersion: OdataVersion.v4,
-                    datasourceType: 'odataServiceUrl' as DatasourceType
+                    datasourceType: 'odataServiceUrl' as DatasourceType,
+                    enableVirtualEndpoints: true
                 },
                 editor,
                 mockVsCode,
-                undefined,
+                {} as Logger,
                 true
             );
 
             const expectedDebugOptions: DebugOptions = {
+                addStartCmd: true,
                 vscode: mockVsCode,
                 sapClientParam: 'sap-client=001',
-                flpAppId: mockProject.flpAppId,
-                flpSandboxAvailable: true,
+                flpAppId: 'app-preview',
+                flpSandboxAvailable: false,
                 isAppStudio: false,
                 odataVersion: '4.0',
                 writeToAppOnly: true
@@ -334,13 +347,15 @@ describe('Test utils', () => {
             const expectedFioriOptions: FioriOptions = {
                 name: mockProject.name,
                 projectRoot: projectPath,
-                debugOptions: expectedDebugOptions
+                debugOptions: expectedDebugOptions,
+                startFile: 'test/flp.html'
             };
-            await createLaunchConfig(projectPath, expectedFioriOptions, editor);
+            expect(createLaunchConfig).toHaveBeenCalledWith(projectPath, expectedFioriOptions, editor, {});
             expect(writeApplicationInfoSettings).toHaveBeenCalledWith(projectPath);
         });
 
         it('should not set odataVersion if service version is not OData v2 or v4', async () => {
+            isAppStudioMock.mockReturnValue(false);
             await generateLaunchConfig(
                 {
                     targetFolder: mockProject.targetFolder,
@@ -350,12 +365,13 @@ describe('Test utils', () => {
                     datasourceType: 'odataServiceUrl' as DatasourceType
                 },
                 editor,
-                mockVsCode
+                mockVsCode,
+                {} as Logger
             );
 
             const expectedDebugOptions: DebugOptions = {
                 vscode: mockVsCode,
-                sapClientParam: '',
+                sapClientParam: 'sap-client=001',
                 addStartCmd: true,
                 flpAppId: mockProject.flpAppId,
                 flpSandboxAvailable: true,
@@ -366,10 +382,11 @@ describe('Test utils', () => {
             const expectedFioriOptions: FioriOptions = {
                 name: mockProject.name,
                 projectRoot: projectPath,
-                debugOptions: expectedDebugOptions
+                debugOptions: expectedDebugOptions,
+                startFile: undefined
             };
 
-            await createLaunchConfig(projectPath, expectedFioriOptions, editor);
+            expect(createLaunchConfig).toHaveBeenCalledWith(projectPath, expectedFioriOptions, editor, {});
             expect(writeApplicationInfoSettings).toHaveBeenCalledWith(projectPath);
         });
     });