+++ examples/odata-cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/odata-cli
 
+## 0.17.8
+
+### Patch Changes
+
+-   Updated dependencies [f9ea9e3]
+    -   @sap-ux/axios-extension@1.22.3
+
+## 0.17.7
+
+### Patch Changes
+
+-   Updated dependencies [14214a3]
+    -   @sap-ux/axios-extension@1.22.2
+
 ## 0.17.6
 
 ### Patch Changes