+++ packages/app-config-writer/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @sap-ux/app-config-writer
 
+## 0.6.15
+
+### Patch Changes
+
+-   Updated dependencies [f9ea9e3]
+    -   @sap-ux/axios-extension@1.22.3
+    -   @sap-ux/project-access@1.30.3
+
+## 0.6.14
+
+### Patch Changes
+
+-   Updated dependencies [14214a3]
+    -   @sap-ux/axios-extension@1.22.2
+
 ## 0.6.13
 
 ### Patch Changes