+++ packages/ui-components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/ui-components
 
+## 1.26.8
+
+### Patch Changes
+
+-   282f30e: feat(ui-components) add icon check
+
 ## 1.26.7
 
 ### Patch Changes