+++ packages/ui5-library-writer/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/ui5-library-writer
 
+## 0.6.3
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.30.3
+
 ## 0.6.2
 
 ### Patch Changes